Bird game basically done, but need to add backrooms element
Changed focus of game. When the player first hits a wall or falls, the bird will get thrown in the backrooms. Right now, when the player touches a wall or the bottom of the screen the player loses control of the bird, but i actually need to add the main backrooms game.